Instead of debian debtags, output a GKeyFile string instead for a given package.
The first commandline argument now has to be the package name for which the plugin .so list follows.

Then e.g this could be called to get the metadata outputted to an .ini-like file of all installed plugins:

for i in list_of_gst_pkgs; do
  qlist -e $i |grep '/usr/lib64/gstreamer-0[.]10/.*[.]so$' |xargs ./gst-codec-info-0.10 $i
done > gst-plugins.ini

Maintainer just has to have all the plugins installed.
A later improvement would be a management tool to instead update the listing on each pkg_postinst on
maintainers machine and notify when the package installing the ini file to users might need to be upgraded.